Making module make-filter
